refactor(core): split feats into loader and grants modules
Extract YAML loading to feats_loader.py and grant resolution to
feats_grants.py; feats.py remains the public facade.
